services/hr_voice_agent/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
httpx==0.27.2
neo4j==5.26.0
asyncpg==0.29.0
pinecone==5.4.2
openai==1.61.1
python-multipart==0.0.9
